refactor(backend): reorganize project modules
Changes:
- Moved `config`, `game`, `mods`, `profiles`, `migrate`, `updater`, and `utils` commands into their feature modules
- Renamed `update/` to `updater/`
- Moved `AppState` and `BundledAssets` into `state.rs`
- Moved log rotation into `utils/logs.rs`
- Moved `get_game_asset` into `utils/misc.rs`
- Simplified `main.rs` to just call the library entry point
- Added `commands` submodules to feature modules
