--- src-tauri/src/commands/mod.rs
@@ -1,7 +1,0 @@
-pub mod config;
-pub mod game;
-pub mod mods;
-pub mod profiles;
-pub mod updater;
-pub mod migrate;
-pub mod utils;
