+++ src-tauri/src/config/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
+pub mod commands;
 
-pub use config::BD2Config;
+pub use config::{BD2Config, AppConfig, PartialAppConfig};
