+++ src-tauri/src/game/mod.rs
@@ -1,2 +1,3 @@
 pub mod game;
 pub mod installer;
+pub mod commands;
