+++ src-tauri/src/migrate/mod.rs
@@ -1,3 +1,4 @@
 pub mod migrate;
+pub mod commands;
 
 pub use migrate::{get_profiles, import_profiles, import_mod_authors};
