+++ src-tauri/src/mods/commands.rs
@@ -1,6 +1,6 @@
 use std::{path::PathBuf};
 
-use bd2modmanager_lib::{
+use crate::{
     ModError, mods::{BD2Mod, install::ModInstallError, preview::{PreviewError, is_texture_mod, preview_image}, sync::SyncMethod, types::BD2ModError}, utils::path::{get_mod_preview_path, get_staging_dir},
 };
 use serde::Serialize;
