+++ src-tauri/src/mods/mod.rs
@@ -7,5 +7,6 @@ pub mod sync;
 pub mod types;
 pub mod utils;
 pub mod preview;
+pub mod commands;
 
 pub use types::{BD2Mod, BD2ModType};
