+++ src-tauri/src/profiles/mod.rs
@@ -1,4 +1,5 @@
 pub mod manager;
 pub mod types;
+pub mod commands;
 
 pub use manager::ProfileManager;
