--- src-tauri/src/update/mod.rs
@@ -1,2 +1,0 @@
-pub mod game_data;
-pub mod mod_preview;
