+++ src-tauri/src/updater/mod.rs
@@ -1,0 +1,3 @@
+pub mod game_data;
+pub mod mod_preview;
+pub mod commands;
