+++ src-tauri/src/utils/mod.rs
@@ -2,5 +2,7 @@ pub mod data;
 pub mod files;
 pub mod misc;
 pub mod path;
+pub mod logs;
+pub mod commands;
 
 pub use files::{has_extension, has_folder, is_archive};
